feat: move logic to stripe elements
