+++ src/Components/PaymentMethod/PaymentMethodContainer.js
@@ -1394,74 +1394,6 @@ const PaymentMethodContainerWithoutStripe = ({
       );
     }
   };
-  useEffect(() => {
-    if (stripe) {
-      const pr = stripe.paymentRequest({
-        country: window.Pelcro.user.location.countryCode || "US",
-        currency: plan.currency,
-        total: {
-          label: plan.nickname || plan.description,
-          amount: state.updatedPrice || plan.amount
-        }
-      });
-      // When Google pay / Apple pay source created
-      pr.on("source", ({ complete, source, ...data }) => {
-        dispatch({ type: DISABLE_COUPON_BUTTON, payload: true });
-        dispatch({ type: DISABLE_SUBMIT, payload: true });
-        dispatch({ type: LOADING, payload: true });
-        complete("success");
-
-        if (source?.card?.three_d_secure === "required") {
-          return generate3DSecureSource(source).then(
-            ({ source, error }) => {
-              if (error) {
-                return handlePaymentError(error);
-              }
-
-              toggleAuthenticationPendingView(true, source);
-            }
-          );
-        }
-
-        dispatch({
-          type: SUBSCRIBE,
-          payload: source
-        });
-      });
-      pr.on("source", ({ complete, source, ...data }) => {
-        dispatch({ type: DISABLE_COUPON_BUTTON, payload: true });
-        dispatch({ type: DISABLE_SUBMIT, payload: true });
-        dispatch({ type: LOADING, payload: true });
-        complete("success");
-
-        if (source?.card?.three_d_secure === "required") {
-          return generate3DSecureSource(source).then(
-            ({ source, error }) => {
-              if (error) {
-                return handlePaymentError(error);
-              }
-
-              toggleAuthenticationPendingView(true, source);
-            }
-          );
-        }
-
-        dispatch({
-          type: SUBSCRIBE,
-          payload: source
-        });
-      });
-      pr.canMakePayment().then((result) => {
-        if (result) {
-          dispatch({
-            type: SET_PAYMENT_REQUEST,
-            payload: pr
-          });
-          dispatch({ type: SET_CAN_MAKE_PAYMENT, payload: !!result });
-        }
-      });
-    }
-  }, [stripe]);
 
   /**
    * Attempt to confirm a Stripe card payment via it's PaymentIntent.
